analytics/ct-logs/requirements.txt
google-cloud-logging
pandas
matplotlib
colour
numpy
mpmath